+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 <!-- insert list items of new changes here -->
 
+* New function `ncattrs` to get attributes from netCDF files hosted on a THREDDS server. 
+
 ## [0.2.0](https://github.com/crim-ca/stac-populator/tree/0.2.0) (2023-11-10)